% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_classes.R
\docType{class}
\name{bam-class}
\alias{bam-class}
\alias{bam}
\title{Class \code{bam} digram}
\description{
Class \code{bam} digram
}
\section{Slots}{

\describe{
\item{\code{sdm_sim}}{A list of sparse vectors representing the area occupied}

\item{\code{palatable_matrices}}{A list of sparse vectors representing palatable sites.}

\item{\code{sim_steps}}{Number of simulation steps
by the species}
}}

\author{
Luis Osorio-Olvera & Jorge Soberón
}
