% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_clusters.R
\name{bam_clusters}
\alias{bam_clusters}
\title{bam_clusters: Function to estimate the connectivity of suitable areas}
\usage{
bam_clusters(model, ngbs = 1, plot_model = FALSE)
}
\arguments{
\item{model}{A niche model in raster format}

\item{ngbs}{Numeric. Number of neighbors (see details).}

\item{plot_model}{Logical. Indicates whether to plot the model in the
cluster map.}
}
\value{
A list with a data.frame of the coordinates of each cluster and a
leaflet map.
}
\description{
Function to estimate the connectivity of suitable areas given
an adjacency matrix.
}
\details{
The grid_base raster object is the area where the dispersal process will
occur.
The number of neighbors depends on the dispersal abilities of the species
and the spatial resolution of the grid_base;
for example, a species's with big dispersal abilities will move throughout
more than 1 km^2 per day, so the idea is to give an approximate number of
moving neighbors (pixels) per unit of time.
}
\examples{

\dontrun{
model_path <- system.file("extdata/Lepus_californicus_cont.tif",
                          package = "bamm")
model <- raster::raster(model_path)
model <- model > 0.7
clusterin <- bamm::bam_clusters(model,ngbs=1,plot_model=T)
clusterin@interactive_map
}
}
