% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_islands.R
\name{bam_islands}
\alias{bam_islands}
\title{bam_islands: Estimates the number of connected patches of a distribution model}
\usage{
bam_islands(model, ngbs = 1)
}
\arguments{
\item{model}{A niche model projected in an area of interest (could be the M).}

\item{ngbs}{Number of pixels the species can disparse.}
}
\value{
A list with the product of AM (A abiotic niche, M accesible area);
the degree matrix of the AM matrix (see)
}
\description{
Estimates the number of connected patches of a distribution model
}
