% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_sim.R
\name{bam_sim}
\alias{bam_sim}
\title{bam_sim: Simulate dispersal dynamics using the set B of the BAM framework.}
\usage{
bam_sim(
  sp1,
  sp2,
  set_M,
  initial_points,
  periods_toxic,
  periods_suitable,
  nsteps,
  progress_bar = TRUE
)
}
\arguments{
\item{sp1}{Niche model of the focal species (the one that disperses).}

\item{sp2}{Niche model of the species with whom sp1 interacts
(currently no dispersal dynamics for this species).}

\item{set_M}{A setM object containing the adjacency matrix for sp1.
See \code{\link[bamm]{adj_mat}}}

\item{initial_points}{A sparse vector returned by the function
\code{\link[bamm]{occs2sparse}}}

\item{periods_toxic}{Time periods that sps2 takes to develop defense
mechanisms (i.e. toxic).}

\item{periods_suitable}{This is the time that sp2 takes to become non-toxic}

\item{nsteps}{Number of steps to run the simulation}

\item{progress_bar}{Show progress bar}
}
\description{
bam_sim: Simulate dispersal dynamics using the set B of the BAM framework.
}
\examples{
\dontrun{
urap <- system.file("extdata/urania_omph/urania_guanahacabibes.tif",
                                  package = "bamm")
ura <- raster::raster(urap)
ompp <- system.file("extdata/urania_omph/omphalea_guanahacabibes.tif",
                                  package = "bamm")
omp <- raster::raster(ompp)
msparse <- bamm::model2sparse(ura)
init_coordsdf <- data.frame(x=-84.38751, y= 22.02932)
initial_points <- bamm::occs2sparse(modelsparse = msparse,init_coordsdf)
set_M <- bamm::adj_mat(modelsparse = msparse,ngbs = 1)
ura_sim <- bamm::bam_sim(sp1=ura, sp2=omp, set_M=set_M,
                        initial_points=initial_points,
                        periods_toxic=3,
                        periods_suitable=3,
                        nsteps=40)
# Animation example
anp <-"C:/Users/l916o895/Dropbox/TeoriadeBAM/articulo_bam/ura_omp_sim.gif"
new_sim <- bamm::sim2Animation(sdm_simul = ura_sim,
                              which_steps = 1:ura_sim@sim_steps,
                              fmt = "GIF",
                              filename = anp)
}
}
