% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_classes.R
\docType{class}
\name{community_sim-class}
\alias{community_sim-class}
\alias{community_bam}
\alias{community-class}
\title{Class  \code{community_sim} digram}
\description{
Class  \code{community_sim} digram
}
\section{Slots}{

\describe{
\item{\code{community_sim}}{A list of sparse vectors representing the area occupied
by the species}
}}

\author{
Luis Osorio-Olvera & Jorge Soberón
}
