% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_classes.R
\docType{class}
\name{csd-class}
\alias{csd-class}
\alias{csd}
\title{Class  \code{csd}}
\description{
Class  \code{csd}
}
\section{Slots}{

\describe{
\item{\code{connections}}{A data.frame with four columns: x, y, clusterID and cluster_size}

\item{\code{interactive_map}}{A leaflet map with markers showing the geographical clusters}

\item{\code{raster_map}}{A raster map with cluster IDs as values.}
}}

\author{
Luis Osorio-Olvera & Jorge Soberón
}
