% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csd_estimate.R
\name{csd_estimate}
\alias{csd_estimate}
\title{csd_estimate: Estimate the connectivity suitability and dispersal plot}
\usage{
csd_estimate(model, dispersal_steps = c(2, 4, 8, 16, 32, 64))
}
\arguments{
\item{model}{A raster model or a setA object representing the suitability model}

\item{dispersal_steps}{A numeric vector with elements representing the dispersal
hypothesis to test.}
}
\description{
csd_plot gives an estimate of the number of geographic clusters
given a set of dispersal hypothesis and a suitability raster
}
\examples{
\dontrun{
model_path <- system.file("extdata/Lepus_californicus_cont.tif",
                          package = "bamm")
model <- raster::raster(model_path)
model <- model > 0.7
csd_plot <- bamm::csd_estimate(model,
                         dispersal_steps=c(2,4,8,16,32,45))


}
}
\author{
Luis Osorio-Olvera & Jorge Soberon
}
