% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csim2pam.R
\name{csim2pam}
\alias{csim2pam}
\title{csim2pam: Converts community simulation to a Presence Absence Matrix (PAM)}
\usage{
csim2pam(community_sim, which_steps)
}
\arguments{
\item{community_sim}{An object of class community_bam.}

\item{which_steps}{Steps in the simulation object to be converted into a PAM}
}
\description{
Converts community simultion object into a
Presence Absence Matrices (PAM) for a given simultation steps.
}
\examples{

\dontrun{
lagos_path <- system.file("extdata/conejos",
                          package = "bamm")
enm_path <- list.files(lagos_path,
                       pattern = ".tif",
                    full.names = TRUE)
en_models <- raster::stack(enm_path)
ngbs_vect <- sample(1:2,replace = T,
                    size = raster::nlayers(en_models))
init_coords <- read.csv(file.path(lagos_path,
                                  "lagos_initit.csv"))
nsteps <- 30
sdm_comm <- bamm::community_sim(en_models = enm_path,
                               ngbs_vect = ngbs_vect,
                               init_coords = init_coords,
                               nsteps = nsteps,
                               threshold = 0.1)

pamt10 <- bamm::csim2pam(community_sim = sdm_comm ,
                       which_steps = 10)
pams <- bamm::csim2pam(community_sim = sdm_comm ,
                      which_steps = c(1:10))
}

}
