% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_sim.R
\name{.nonzero}
\alias{.nonzero}
\title{Helper function to compute the elements in g0
that have no zero values.The function is taken from the
Ringo package}
\usage{
.nonzero(x)
}
\arguments{
\item{x}{A matrix of class "dgCMatrix"}
}
\description{
Helper function to compute the elements in g0
that have no zero values.The function is taken from the
Ringo package
}
