% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_bam.R
\name{eigen_bam}
\alias{eigen_bam}
\title{eigen_bam: Compute the Eigen system of two bam objects}
\usage{
eigen_bam(A = NULL, M = NULL, which_eigen = 1, rmap = TRUE)
}
\arguments{
\item{A}{A bam object of class setA.}

\item{M}{A bam object of class setM.}

\item{which_eigen}{An integer representing the which eigen value and eigen
vector will be computed.}

\item{rmap}{Logical. If TRUE the function will return a map of the eigen
vector of the product AM.}
}
\value{
A table
}
\description{
Calucultates the Eigen values and Eigen vectors of bam objects
}
\examples{
\dontrun{
model_path <- system.file("extdata/Lepus_californicus_cont.tif",
                          package = "bamm")
model <- raster::raster(model_path)
sparse_mod <- bamm::model2sparse(model = model)
adj_mod <- bamm::adj_mat(sparse_mod,ngbs = 1,eigen_sys = T)
eig_bam <- bamm::eigen_bam(A=sparse_mod,M=adj_mod)
raster::plot(eig_bam$map)
}
}
