% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_classes.R
\docType{class}
\name{leaflet-class}
\alias{leaflet-class}
\alias{leaflet}
\title{Class  \code{leaflet} leaflet}
\value{
An object of class leaflet
}
\description{
Class  \code{leaflet} leaflet
}
\author{
Luis Osorio-Olvera & Jorge Soberón
}
