% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_methods.R
\name{plot,diversity_range,ANY-method}
\alias{plot,diversity_range,ANY-method}
\title{Plot method for objects of class diversity_range \pkg{bamm}.}
\usage{
\S4method{plot}{diversity_range,ANY}(
  x,
  xlab = NULL,
  plot_type = "diversity_range",
  legend = TRUE,
  legend_position = "bottomright",
  ylab = NULL,
  col = NULL,
  pch = NULL,
  pch_legend = 19,
  radius = 0.5,
  ...
)
}
\arguments{
\item{x}{An object of class diversity_range}

\item{xlab}{x label}

\item{plot_type}{Plot type: possible options: "diversity_range"
(range-diversity plot),
                 "diversity_range_map" (a raster map with
                 diversity_range categories),
                 "alpha" (a raster map with alpha diversity values),
                 "dispersion_field" (a raster with dispersion field)}

\item{legend}{Logical. If TRUE the legend of the categorical diversity
range values will appear.}

\item{legend_position}{Legend position.}

\item{ylab}{y label}

\item{col}{Plot colors.}

\item{pch}{Patch type.}

\item{pch_legend}{Patch type for legends.}

\item{radius}{Size of the patch for the interactive map.}

\item{...}{Graphical parameters. Any argument that can be passed to 1)
base::plot, such as axes=FALSE, main='title', ylab='latitude' 2)
leaflet::leaflet or 3) leaflet::addCircleMarkers.}
}
\value{
Plot of the results of the diversity_range analysis
}
\description{
Plot method for objects of class diversity_range \pkg{bamm}.
}
\details{
To show interactive diversity_range plots install the 'plotly' R
package.
}
\author{
Luis Osorio-Olvera & Jorge Soberón
}
