% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_classes.R
\docType{class}
\name{setM-class}
\alias{setM-class}
\alias{setM}
\alias{M-class}
\title{Class for the M set of the \code{bamm} diagram}
\value{
An object of class setM
}
\description{
Class for the M set of the \code{bamm} diagram
}
\section{Slots}{

\describe{
\item{\code{adj_matix}}{An adjacency matrix}

\item{\code{adj_list}}{An adjacency list}

\item{\code{initial_points}}{A presence-absence vector with species' occurrences}

\item{\code{n_initial_points}}{Number of initial points used to start the dispersal
process}

\item{\code{ngbs}}{Number of neighbors}
}}

\examples{
showClass("setM")
}
\author{
Luis Osorio-Olvera & Jorge Soberón
}
