% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape2Grid.R
\name{shape2Grid}
\alias{shape2Grid}
\title{shape2Grid: Function to create a grid given a spatial polygon}
\usage{
shape2Grid(shpolygon, resolution, ones = TRUE)
}
\arguments{
\item{shpolygon}{A SpatialPolygon, SpatialPolygonDataFrame representing the desired shape of the grid.}

\item{resolution}{Numeric. Spatial resoltion of the grid.}

\item{ones}{Logical. Fill with ones the values of the raster. If not the values will be written as cellID values.}
}
\value{
Returns a raster object whith the shape of 'shpolygon' of a given resoltion.
}
\description{
shapeToGrid creates a raster grid given a spatial polygon and a grid resolution.
}
\examples{
\dontrun{
data("wrld_simpl", package = "maptools")
mx <- wrld_simpl[wrld_simpl$NAME=="Mexico",]
mx_grid <- bamm::shape2Grid(mx,0.5)
}
}
