% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_methods.R
\name{show,setA-method}
\alias{show,setA-method}
\alias{show,csd-method}
\alias{show,pam-method}
\alias{show,bioindex_sparse-method}
\alias{show,setM-method}
\alias{show,diversity_range-method}
\title{Show information in setA class \pkg{bamm}.}
\usage{
\S4method{show}{setA}(object)

\S4method{show}{csd}(object)

\S4method{show}{pam}(object)

\S4method{show}{bioindex_sparse}(object)

\S4method{show}{setM}(object)

\S4method{show}{diversity_range}(object)
}
\arguments{
\item{object}{An object of class diversity_range}
}
\value{
Display information about the setA object

Display information about the csd object

Display information about the pam object

Display information about the bioindex_spars object

Display information about the setM object

Display information about the diversity_range object
}
\description{
Show information in setA class \pkg{bamm}.

Show information in csd class \pkg{bamm}.

Show information in pam class \pkg{bamm}.

Show information in pam class \pkg{bamm}.

Show information in setA class \pkg{bamm}.

Show information in diversity_range class \pkg{bamm}.
}
\author{
Luis Osorio-Olvera & Jorge Soberón
}
