% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_classes.R
\docType{class}
\name{bioindex-class}
\alias{bioindex-class}
\alias{bioindex}
\alias{biodiversity_index}
\title{Class  \code{bioindex}}
\description{
Class  \code{bioindex}
}
\section{Slots}{

\describe{
\item{\code{alpha}}{A matrix with the richness of species per site}

\item{\code{omega}}{A matrix with the range size of every species}

\item{\code{dispersion_field}}{A matrix with the set of ranges of all species that occur in at each locality}
}}

\author{
Luis Osorio-Olvera & Jorge Soberón
}
