% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_sim.R
\name{community_sim}
\alias{community_sim}
\title{community_bam: Community \code{bamm}}
\usage{
community_sim(en_models, ngbs_vect, init_coords, nsteps, parallel, ...)
}
\arguments{
\item{en_models}{A stack or directory with the ecological niche models for
each species in the community.}

\item{ngbs_vect}{A vector}

\item{init_coords}{A data.frame with 3 columns: sp_name, x and
y; x is the longitude and y is the latitude of initial dispersal points}

\item{nsteps}{Number of iteration steps for the simulation.}

\item{parallel}{TRUE if the simulation should be run in parallel.
Default FALSE}

\item{...}{Parameters of the function \code{\link[bamm]{model2sparse}}}
}
\description{
Estimate community dynamics using the \code{bamm} framework
}
\examples{
\dontrun{
lagos_path <- system.file("extdata/conejos",
                          package = "bamm")
enm_path <- list.files(lagos_path,
                       pattern = ".tif",
                       full.names = TRUE)[seq(1,10)]
en_models <- raster::stack(enm_path)
ngbs_vect <- sample(1:2,replace = TRUE,
                    size = raster::nlayers(en_models))
init_coords <- read.csv(file.path(lagos_path,
                                  "lagos_initit.csv"))[seq(1,10),]
nsteps <- 12
sdm_comm <- bamm::community_sim(en_models = enm_path,
                               ngbs_vect = ngbs_vect,
                               init_coords = init_coords,
                               nsteps = nsteps,
                               threshold = 0.1)

com_pam <- bamm::csim2pam(sdm_comm,which_steps = seq(1,nsteps))
rich_pam <- pam2richness(com_pam,which_steps = c(1,5,10))
raster::plot(rich_pam)
}
}
