% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_classes.R
\docType{class}
\name{g_area-class}
\alias{g_area-class}
\alias{g_area}
\alias{coordinates-class}
\title{S4 classes to organize data and results of \code{bamm} objects}
\value{
An object of class g_area
}
\description{
S4 classes to organize data and results of \code{bamm} objects
}
\section{Slots}{

\describe{
\item{\code{coordinates}}{A two column matrix with coordinates}

\item{\code{eigen_vec}}{Eigen vector of adjacency matrix}

\item{\code{eigen_val}}{Eigen value of adjacency matrix
slot g_model A raster representing the geographic area
slot g_sparse A sparse matrix of the geographic area}
}}

\author{
Luis Osorio-Olvera & Jorge Soberón
}
