% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_classes.R
\docType{class}
\name{pam-class}
\alias{pam-class}
\alias{pam}
\alias{PAM}
\title{Class  \code{pam} Presence-Absence Matrix}
\description{
Class  \code{pam} Presence-Absence Matrix
}
\section{Slots}{

\describe{
\item{\code{pams}}{A list of sparse matices representing Presence-Absecen Matrix for
each simulation time}

\item{\code{which_steps}}{Simulation steps}

\item{\code{sp_names}}{Names of species in the PAM}

\item{\code{grid}}{Raster grid of the studied area}

\item{\code{cellIDs}}{Cells with ids of the PAM sites}
}}

\author{
Luis Osorio-Olvera & Jorge Soberón
}
