% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_classes.R
\docType{class}
\name{setA-class}
\alias{setA-class}
\alias{setA}
\alias{A-class}
\alias{showClass("setA")}
\title{Class for the A set of the BAM diagram}
\description{
A class for the A set of the BAM diagram. It contains raster models and
IDs of pixels with values different than NA.
}
\section{Slots}{

\describe{
\item{\code{niche_model}}{A niche model in raster format. It can be a binary model or continuous.
If the model is in a continuous format.}

\item{\code{suit_threshold}}{Suitability value used to binarize continuous model}

\item{\code{cellIDs}}{A numeric vector with the IDs of the cells with prediction values}

\item{\code{suit_values}}{A numeric vector with suitability value of the continuous map}

\item{\code{sparse_model}}{A niche model in sparse matrix format}
}}

\author{
Luis Osorio-Olvera & Jorge Soberón
}
