% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape2Grid.R
\name{shape2Grid}
\alias{shape2Grid}
\title{shape2Grid: Function to create a grid given a spatial polygon}
\usage{
shape2Grid(shpolygon, resolution, ones = TRUE)
}
\arguments{
\item{shpolygon}{A SpatialPolygon, SpatialPolygonDataFrame representing
the desired shape of the grid.}

\item{resolution}{Numeric. Spatial resolution of the grid.}

\item{ones}{Logical. Fill with ones the values of the raster. If not the
values will be written as cellID values.}
}
\value{
Returns a raster object with the shape of 'shpolygon' of a given
 resolution.
}
\description{
shapeToGrid creates a raster grid given a spatial polygon and
a grid resolution.
}
\examples{
\dontrun{
x_coord <- c(16.48438,  17.49512,  24.74609, 22.59277, 16.48438)
y_coord <- c(59.736328125, 55.1220703125, 55.0341796875, 61.142578125,
             59.736328125)
xy <- cbind(x_coord, y_coord)
p <- sp::Polygon(xy)
ps <- sp::Polygons(list(p),1)
sps <- sp::SpatialPolygons(list(ps))
r1 <- bamm::shape2Grid(sps,resolution = 0.5,ones = T)
}
}
